feat: Brought back benchmarking.
The elimination of the GitOid crate had broken benchmarking, and it's now
fixed again.
